flask-multi-db-monorepo/apps/postgresql-app/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.3
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
python-dotenv==0.19.2
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1